[refactor] split handler, service
